Added new ground and platform textures
Exactly what it says on the tin...
NB: Some sprites' pixels per unit were changed. (I'm trying to get
everything to the same size (currently using 50 ppu, but when averything
is the same size this can be antything we want). so, if I changed
anything and now it doesn't fit or something, let me know)